test/ciknow/dao/hibernate/NodeHibernateDaoTest.java
package ciknow.dao.hibernate;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.HashSet;
import java.util.Iterator;
import java.util.List;
import java.util.Map;
import java.util.Set;

import junit.framework.Assert;

import org.apache.log4j.Logger;
import org.junit.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.TestExecutionListeners;

import ciknow.dao.EdgeDao;
import ciknow.dao.GroupDao;
import ciknow.dao.NodeDao;
import ciknow.dao.QuestionDao;
import ciknow.dao.RoleDao;
import ciknow.dao.SurveyDao;
import ciknow.domain.Edge;
import ciknow.domain.Group;
import ciknow.domain.Node;
import ciknow.domain.Question;
import ciknow.domain.Role;
import dbunit.Dbunit;
import dbunit.DbunitTestExecutionListener;


@TestExecutionListeners(DbunitTestExecutionListener.class)
//@org.springframework.test.context.transaction.TransactionConfiguration(transactionManager="transactionManager", defaultRollback=false)
public class NodeHibernateDaoTest extends AbstractHibernateDaoTest {
	
	private static final int MAX_NODE_ROW_COUNT = 10;
	
	@Autowired
	private NodeDao nodeDao;
	@Autowired
	private RoleDao roleDao;
	@Autowired
	private GroupDao groupDao;
	@Autowired
	private EdgeDao edgeDao;
	@Autowired
	private SurveyDao surveyDao;
	@Autowired private QuestionDao questionDao;
	
	private static final Logger logger = Logger
			.getLogger(NodeHibernateDaoTest.class.getName());
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void save() {
		
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodeDao.getCount());
		
		HashSet<Role> roles = new HashSet<Role>();
		Role r = new Role();
		r.setName("Test_Role");
		r.setVersion(1L);
		roles.add(r);
		r = new Role();
		r.setName("Test_Role2");
		roles.add(r);
		HashMap<String, String> attributes = new HashMap<String, String>();
		attributes.put("1", "ATT_VALUE 1");
		attributes.put("2", "ATT_VALUE 2");
		HashMap<String, String> longAttributes = new HashMap<String, String>();
		longAttributes.put("1", "LONG_ATT_VALUE 1");
		longAttributes.put("2", "LONG_ATT_VALUE 2");
		Group group = new Group();
		group.setName("New Group");
		group.setVersion(1L);
		Group group2 = new Group();
		group2.setName("Newer Group");
		group2.setVersion(1L);
		HashSet<Group> groups = new HashSet<Group>();
		groups.add(group);
		groups.add(group2);
		
	
		Node node = createNode(1L, "John", "Middle", "Johnson",
				"username", "password", "nodeType", 
				"JOHN_LABEL", true, "http://johnJohnson",
				"john@johnson.com", "888-777-2222", "888-777-1111", "888-777-3333",
				"address1", "address2", "Chicago", "Illinois", "64111","USA", 
				"department", "Organization", "Unit", roles, attributes, longAttributes, groups);
		
		nodeDao.save(node);
		
		
		Assert.assertEquals(MAX_NODE_ROW_COUNT + 1, nodeDao.getCount());
		
		List<Node> nodes = nodeDao.getAll();
		compareNodes(node, nodes.get(nodes.size()-1));
		
		
		
	}
	
	

	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void saveCollection(){
		
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodeDao.getCount());
		
		ArrayList<Node> nodes = new ArrayList<Node>();
		
		HashSet<Role> roles = new HashSet<Role>();
		Role r = new Role();
		r.setName("Test_Role");
		r.setVersion(1L);
		roles.add(r);
		r = new Role();
		r.setName("Test_Role2");
		roles.add(r);
		HashMap<String, String> attributes = new HashMap<String, String>();
		attributes.put("1", "ATT_VALUE 1");
		attributes.put("2", "ATT_VALUE 2");
		HashMap<String, String> longAttributes = new HashMap<String, String>();
		longAttributes.put("1", "LONG_ATT_VALUE 1");
		longAttributes.put("2", "LONG_ATT_VALUE 2");
		Group group = new Group();
		group.setName("New Group");
		group.setVersion(1L);
		Group group2 = new Group();
		group2.setName("Newer Group");
		group2.setVersion(1L);
		HashSet<Group> groups = new HashSet<Group>();
		groups.add(group);
		groups.add(group2);
		
	
		Node node = createNode(1L, "John", "Middle", "Johnson",
				"username", "password", "nodeType", 
				"JOHN_LABEL", true, "http://johnJohnson",
				"john@johnson.com", "888-777-2222", "888-777-1111", "888-777-3333",
				"address1", "address2", "Chicago", "Illinois", "64111","USA", 
				"department", "Organization", "Unit", roles, attributes, longAttributes, groups);
		nodes.add(node);		
		Node node2 = createNode(1L, "John2", "Middle2", "Johnson2",
				"username", "password", "nodeType", 
				"JOHN_LABEL", true, "http://johnJohnson",
				"john@johnson.com", "888-777-2222", "888-777-1111", "888-777-3333",
				"address1", "address2", "Chicago", "Illinois", "64111","USA", 
				"department", "Organization", "Unit", roles, attributes, longAttributes, groups);
		nodes.add(node2);
		nodeDao.save(nodes);
		
		Assert.assertEquals(MAX_NODE_ROW_COUNT + 2, nodeDao.getCount());
		
		List<Node> nodes2 = nodeDao.getAll();
		compareNodes(node, nodes2.get(nodes2.size()-2));
		compareNodes(node2, nodes2.get(nodes2.size()-1));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void delete(){
		
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodeDao.getCount());
		Node node = nodeDao.loadById(1L);
		clearConstraints(node);
		nodeDao.delete(node);
		Assert.assertEquals(MAX_NODE_ROW_COUNT-1, nodeDao.getCount());
		
		Assert.assertTrue(!nodeDao.getAll().contains(node));
		
	}
	
	
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void deleteCollection(){
		
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodeDao.getCount());
		Node node = nodeDao.loadById(10L);
		clearConstraints(node);
		Node node2 = nodeDao.loadById(9L);
		clearConstraints(node2);
		ArrayList<Node> nodes = new ArrayList<Node>();
		nodes.add(node);
		nodes.add(node2);
		nodeDao.delete(nodes);
		Assert.assertEquals(MAX_NODE_ROW_COUNT-2, nodeDao.getCount());
		
		Assert.assertTrue(!nodeDao.getAll().contains(node));
		Assert.assertTrue(!nodeDao.getAll().contains(node2));
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void deleteAll(){
		
		/*Assert.assertEquals(MAX_NODE_ROW_COUNT, nodeDao.getCount());
		List<Node> nodes = nodeDao.getAll();
		for(Node n: nodes){
			clearConstraints(n);
		}
		nodeDao.deleteAll();
		Assert.assertEquals(0, nodeDao.getCount());
		*/
		
		//TODO address bulk update issue
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getAll(){
		
		List<Node> nodes = nodeDao.getAll();
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
		
		node = nodes.get(5);
		Assert.assertEquals(new Long(6L), node.getId());
		
		node = nodes.get(9);
		Assert.assertEquals(new Long(10L), node.getId());
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void loadAll(){
		
		List<Node> nodes = nodeDao.loadAll();
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
		Set<Role> roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(1L)));
		Set<Group> groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(2L)));
		
		node = nodes.get(5);
		Assert.assertEquals(new Long(6L), node.getId());
		
		node = nodes.get(9);
		Assert.assertEquals(new Long(10L), node.getId());
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getAllPlainNodes(){
		
		Map<String, Map> plainNodes = nodeDao.getAllPlainNodes();
		Map map = plainNodes.get("1");
		Assert.assertEquals("admin", map.get("username"));
		Assert.assertEquals(new Long(1), map.get("nodeId"));
		Assert.assertEquals("CIKNOW ADMIN", map.get("label"));
		
		map = plainNodes.get("5");
		Assert.assertEquals("node_5", map.get("username"));
		Assert.assertEquals(new Long(5), map.get("nodeId"));
		Assert.assertEquals("NODE_5_TEST", map.get("label"));
		
		map = plainNodes.get("10");
		Assert.assertEquals("node_10", map.get("username"));
		Assert.assertEquals(new Long(10), map.get("nodeId"));
		Assert.assertEquals("NODE_10_TEST", map.get("label"));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getPlainNodesByIds(){
		//List<String> attributes, List<Long> nodeIds
		
		//TODO address failure in both dbunit test case (no records found) 
		// and in dao test (attribute index out of bounds error - dao code fixed)
		
		
		ArrayList<String> attributes = new ArrayList<String>();
		/*attributes.add("version");
		attributes.add("label");
		attributes.add("type");
		
		ArrayList<Long> nodeIds = new ArrayList<Long>();
		nodeIds.add(1L);
		//nodeIds.add(3L);
		
		Map<String, Map> plainNodes = nodeDao.getPlainNodesByIds(attributes, nodeIds);
		System.out.println("plainNodes.size: " + plainNodes.size());
		
		Iterator<String> it = plainNodes.keySet().iterator();
		while(it.hasNext()){
			String key = it.next();
			System.out.println("NEXT KEY: " + key);
			Map map = plainNodes.get(key);
			Iterator it2 = map.keySet().iterator();
			while(it2.hasNext()){
				String key2 = (String)it2.next();
				System.out.println("key, value: " + key2 + ", " + map.get(key2));
			}
		}
		*/
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getCount(){
		Assert.assertEquals(MAX_NODE_ROW_COUNT, nodeDao.getCount());
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getNodeTypes(){
		
		List<String> nodeTypes = nodeDao.getNodeTypes();
		Assert.assertEquals(6, nodeTypes.size());
		Assert.assertTrue(nodeTypes.contains("user"));
		Assert.assertTrue(nodeTypes.contains("user2"));
		Assert.assertTrue(nodeTypes.contains("user3"));
		Assert.assertTrue(nodeTypes.contains("user4"));
		Assert.assertTrue(nodeTypes.contains("directorate"));
		Assert.assertTrue(nodeTypes.contains("tag"));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getProxy(){
		
		Node node = nodeDao.getProxy(1L);
		Assert.assertEquals(new Long(1), node.getId());
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findById(){
		Node node = nodeDao.findById(1L);
		Assert.assertEquals(new Long(1), node.getId());
		Assert.assertEquals("CIKNOW ADMIN", node.getLabel());
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void loadById(){
		Node node = nodeDao.loadById(1L);
		Assert.assertEquals(new Long(1L), node.getId());
		Set<Role> roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(1L)));
		Set<Group> groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(2L)));
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findByIds(){
		
		ArrayList<Long> ids = new ArrayList<Long>();
		ids.add(1L);
		ids.add(2L);
		List<Node> nodes = nodeDao.findByIds(ids);
		Assert.assertEquals(2, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
		
		node = nodes.get(1);
		Assert.assertEquals(new Long(2L), node.getId());
		
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void loadByIds(){
		
		ArrayList<Long> ids = new ArrayList<Long>();
		ids.add(1L);
		ids.add(2L);
		List<Node> nodes = nodeDao.loadByIds(ids);
		Assert.assertEquals(2, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
		Set<Role> roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(1L)));
		Set<Group> groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(2L)));
		
		node = nodes.get(1);
		Assert.assertEquals(new Long(2L), node.getId());
		roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(2L)));
		groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(1L)));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
    public void findByType(){
    	
		List<Node> nodes = nodeDao.findByType("user2");
		Assert.assertEquals(2, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(4L), node.getId());
		
		node = nodes.get(1);
		Assert.assertEquals(new Long(6L), node.getId());
		
	
    }
    
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
    public void loadByType(){
		
		List<Node> nodes = nodeDao.findByType("user");
		Assert.assertEquals(3, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
		Assert.assertEquals(new Long(1L), node.getId());
		Set<Role> roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(1L)));
		Set<Group> groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(2L)));
		
		node = nodes.get(1);
		Assert.assertEquals(new Long(3L), node.getId());
		
		node = nodes.get(2);
		Assert.assertEquals(new Long(5L), node.getId());
		
		
		
    }
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
    public void findByUsername(){
		
		Node node = nodeDao.findByUsername("admin");
	
		Assert.assertEquals(new Long(1L), node.getId());
    	
    }
    
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void loadByUsername(){
		
		Node node = nodeDao.loadByUsername("admin");
		
		Assert.assertEquals(new Long(1L), node.getId());
		Set<Role> roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(1L)));
		Set<Group> groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(2L)));
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findByUsernames(){
		
		ArrayList<String> usernames = new ArrayList<String>();
		usernames.add("node_4");
		usernames.add("node_5");
		List<Node> nodes = nodeDao.findByUsernames(usernames);
		Assert.assertEquals(2, nodes.size());
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(4L), node.getId());
		
		node = nodes.get(1);
		Assert.assertEquals(new Long(5L), node.getId());
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findByLabel(){
		
		List<Node> nodes = nodeDao.findByLabel("CIKNOW ADMIN");
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
	
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void loadByLabel(){
		
		List<Node> nodes = nodeDao.loadByLabel("CIKNOW ADMIN");
		
		Node node = nodes.get(0);
		Assert.assertEquals(new Long(1L), node.getId());
		Set<Role> roles = node.getRoles();
		Assert.assertTrue(roles.contains(roleDao.loadById(1L)));
		Set<Group> groups = node.getGroups();
		Assert.assertTrue(groups.contains(groupDao.loadById(2L)));
		
		
	}
	
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findByQuestion(){
		
		Question question = questionDao.findById(3L);
		List<Node> nodes = nodeDao.findTagsByQuestion(question);
		Assert.assertEquals(1, nodes.size());
		
		Assert.assertEquals(new Long(7), nodes.get(0).getId());
	}
	
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findByAttribute(){
		
		//method not implemented
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void matchAttribute(){
		
		//method not implemented
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
    public void findTagByQuestion(){
		
		/*
		Question q = questionDao.findById(1L);
		List<Node> nodes = nodeDao.findTagByQuestion(q);
		
		System.out.println("nodes.size: " + nodes.size());
    	*/
		
		//TODO investigate test case failure. SQL insert times out. 
		//Note - test doesn't fail when executed from NodeHibernateDao.main
    }

	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findConnectedCollection(){
		
		Node node1 = nodeDao.loadById(1L);
		Node node2 = nodeDao.loadById(2L);
		ArrayList<Node> list = new ArrayList<Node>();
		list.add(node2);
		list.add(node1);
		
		List<Node> nodes = nodeDao.findConnected(list, true);
		Assert.assertEquals(4, nodes.size());
		Assert.assertEquals(new Long(9), nodes.get(0).getId());
		Assert.assertEquals(new Long(1), nodes.get(1).getId());
		Assert.assertEquals(new Long(6), nodes.get(2).getId());
		Assert.assertEquals(new Long(7), nodes.get(3).getId());
		
	}
	

	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findConnectingCollection(){
		
		Node node7 = nodeDao.loadById(7L);
		Node node3 = nodeDao.loadById(3L);
		
		ArrayList<Node> list = new ArrayList<Node>();
		list.add(node7);
		list.add(node3);
		List<Node> nodes = nodeDao.findConnecting(list, true);
		
		Assert.assertEquals(3, nodes.size());
		Assert.assertEquals(new Long(4), nodes.get(0).getId());
		Assert.assertEquals(new Long(8), nodes.get(1).getId());
		Assert.assertEquals(new Long(2), nodes.get(2).getId());
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void findNeighbors(){
		
		Node node5 = nodeDao.loadById(5L);
		Node node7 = nodeDao.loadById(7L);
		
		ArrayList<Node> list = new ArrayList<Node>();
		list.add(node5);
		list.add(node7);
		
		ArrayList<String> edgeTypes = new ArrayList<String>();
		edgeTypes.add("Edge_Type3");
		edgeTypes.add("Edge_Type2");
		
		Set<Node> nodes = nodeDao.findNeighbors(list, true, true, edgeTypes);
		Node node2 = nodeDao.loadById(2L);
		Node node4 = nodeDao.loadById(4L);
		Node node6 = nodeDao.loadById(6L);
		Node node8 = nodeDao.loadById(8L);
	
		Assert.assertEquals(4, nodes.size());
		Assert.assertTrue(nodes.contains(node2));
		Assert.assertTrue(nodes.contains(node4));
		Assert.assertTrue(nodes.contains(node6));
		Assert.assertTrue(nodes.contains(node8));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void deleteAttributeByKey(){
		
		Node node1 = nodeDao.loadById(1L);
		Node node2 = nodeDao.loadById(2L);
		Assert.assertEquals("NODE_1_ATT_VALUE_1", node1.getAttribute("1"));
		Assert.assertEquals("NODE_2_ATT_VALUE_1", node2.getAttribute("1"));
		
		nodeDao.deleteAttributeByKey("1");
		
		nodeDao.getAll();
		
		node1 = nodeDao.loadById(1L);
		node2 = nodeDao.loadById(2L);
		System.out.println("ATT: " + node1.getAttribute("1"));
		
		//TODO investigate test case failure. 
		//Note - test doesn't fail when executed from NodeHibernateDao.main
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void deleteLongAttributeByKey(){
		
		Node node1 = nodeDao.loadById(1L);
		Node node2 = nodeDao.loadById(2L);
		Assert.assertEquals("LONG_NODE_1_ATT_VALUE_1", node1.getLongAttribute("1"));
		Assert.assertEquals("LONG_NODE_2_ATT_VALUE_1", node2.getLongAttribute("1"));
		
		nodeDao.deleteAttributeByKey("1");
		
		nodeDao.getAll();
		
		node1 = nodeDao.loadById(1L);
		node2 = nodeDao.loadById(2L);
		System.out.println("ATT: " + node1.getLongAttribute("1"));
		
		//TODO investigate test case failure. Same condition as deleteAttributeByKey 
		
	}
	
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getAttributeNames(){
		
		ArrayList<String> nodeTypes = new ArrayList<String>();
		nodeTypes.add("user");
		nodeTypes.add("user2");
		List<String> attNames = nodeDao.getAttributeNames(nodeTypes);
	
		Assert.assertTrue(attNames.contains("1"));
		Assert.assertTrue(attNames.contains("2"));
		Assert.assertTrue(attNames.contains("3"));
		Assert.assertTrue(attNames.contains("4"));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getLongAttributeNames(){
		
		List<String> attNames = nodeDao.getLongAttributeNames();
		
		Assert.assertTrue(attNames.contains("1"));
		Assert.assertTrue(attNames.contains("2"));
		Assert.assertTrue(attNames.contains("3"));
	
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getAttributeValues(){
		
		List<String> attValues = nodeDao.getAttributeValues("1");

		Assert.assertTrue(attValues.contains("NODE_1_ATT_VALUE_1"));
		Assert.assertTrue(attValues.contains("NODE_2_ATT_VALUE_1"));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getPropertyValues(){
		
		List<String> properties = nodeDao.getPropertyValues("type");
		
		Assert.assertEquals(6, properties.size());
		Assert.assertTrue(properties.contains("user"));
		Assert.assertTrue(properties.contains("user2"));
		Assert.assertTrue(properties.contains("user3"));
		Assert.assertTrue(properties.contains("user4"));
		Assert.assertTrue(properties.contains("directorate"));
		Assert.assertTrue(properties.contains("tag"));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getNodeIdsWithAttribute(){

		List<Long> ids = nodeDao.getNodeIdsWithAttribute("1");
		
		Assert.assertEquals(2, ids.size());
		Assert.assertTrue(ids.contains(new Long(1)));
		Assert.assertTrue(ids.contains(new Long(2)));
		
	}
	
	@Test
	@Dbunit(dataSetLocation="classpath:/dbunit/nodes.xml")
	public void getScoreMapByAttrName(){

		//Map<Long, Float> map = nodeDao.getScoreMapByAttrName("5`Q_2");
		
		//TODO investigate test case failure. 
		////Note - test doesn't fail when executed from NodeHibernateDao.main
	}
	
	
	private void clearConstraints(Node node){
		
		List<Group> groups = groupDao.getAll();
		for(int i=0; i<groups.size(); i++){
			node.getGroups().remove(groups.get(i));
		}
		
		List<Role> roles = roleDao.getAll();
		for(int i=0; i<roles.size(); i++){
			node.getRoles().remove(roles.get(i));
		}
	
		edgeDao.delete(edgeDao.findByFromNodeId(node.getId()));
        edgeDao.delete(edgeDao.findByToNodeId(node.getId()));
        List<Edge> edges = edgeDao.findByCreatorId(node.getId());
        for(int i=0; i<edges.size(); i++){
        	edgeDao.delete(edges.get(i));
        }
        
        surveyDao.delete(surveyDao.findByDesigner(node));
        
        nodeDao.save(node);
		
	}
	
	
	private void compareNodes(Node expected, Node actual){
		
		Assert.assertEquals(expected.getId(), actual.getId());
		Assert.assertEquals(expected.getAddr1(), actual.getAddr1());
		Assert.assertEquals(expected.getAddr2(), actual.getAddr2());
		Assert.assertEquals(expected.getCell(), actual.getCell());
		Assert.assertEquals(expected.getCity(), actual.getCity());
		Assert.assertEquals(expected.getCountry(), actual.getCountry());
		Assert.assertEquals(expected.getDepartment(), actual.getDepartment());
		Assert.assertEquals(expected.getEmail(), actual.getEmail());
		Assert.assertEquals(expected.getEnabled(), actual.getEnabled());
		Assert.assertEquals(expected.getFax(), actual.getFax());
		Assert.assertEquals(expected.getFirstName(), actual.getFirstName());
		Assert.assertEquals(expected.getLabel(), actual.getLabel());
		Assert.assertEquals(expected.getLastName(), actual.getLastName());
		Assert.assertEquals(expected.getMidName(), actual.getMidName());
		Assert.assertEquals(expected.getOrganization(), actual.getOrganization());
		Assert.assertEquals(expected.getPassword(), actual.getPassword());
		Assert.assertEquals(expected.getPhone(), actual.getPhone());
		Assert.assertEquals(expected.getState(), actual.getState());
		Assert.assertEquals(expected.getType(), actual.getType());
		Assert.assertEquals(expected.getUnit(), actual.getUnit());
		Assert.assertEquals(expected.getUri(), actual.getUri());
		Assert.assertEquals(expected.getUsername(), actual.getUsername());
		Assert.assertEquals(expected.getZipcode(), actual.getZipcode());
		Assert.assertEquals(expected.getRoles(), actual.getRoles());
		Assert.assertEquals(expected.getGroups(), actual.getGroups());
		
		Iterator<String> keys = expected.getAttributes().keySet().iterator();
		while(keys.hasNext()){
			String key = keys.next();
			Assert.assertEquals(expected.getAttribute(key), actual.getAttribute(key));
		}		
		keys = expected.getLongAttributes().keySet().iterator();
		while(keys.hasNext()){
			String key = keys.next();
			Assert.assertEquals(expected.getLongAttribute(key), actual.getLongAttribute(key));
		}
		
	}
	
	private Node createNode(Long version, String firstName, String middleName, String lastName,
			String username, String password,  String type, 
			String label, boolean enabled, String uri,
			String email, String phone, String cell, String fax,
			String address1, String address2, String city, String state, String zipcode, String country, 
			String department, String organization, String unit, Set<Role> roles,
			HashMap<String, String> attributes, HashMap<String, String> longAttributes, Set<Group> groups){
		
		Node node = new Node();
		
		node.setAddr1(address1);
		node.setAddr2(address2);
		node.setCell(cell);
		node.setCity(city);
		node.setState(state);
		node.setZipcode(zipcode);
		node.setCountry(country);
		node.setDepartment(department);
		node.setEmail(email);
		node.setEnabled(enabled);
		node.setFax(fax);
		node.setFirstName(firstName);
		node.setMidName(middleName);
		node.setLastName(lastName);
		node.setLabel(label);
		node.setOrganization(organization);
		node.setPassword(password);
		node.setPhone(phone);
		node.setType(type);
		node.setUnit(unit);
		node.setUri(uri);
		node.setUsername(username);
		node.setVersion(version);
		node.setRoles(roles);
		node.setAttributes(attributes);
		node.setLongAttributes(longAttributes);
		node.setGroups(groups);
		
		
		return node;
		
	}
	
}
